+++ src/planetaryprotector/research/horizontal/noun/NounMine.java
@@ -3,15 +3,15 @@ import planetaryprotector.Core;
 import planetaryprotector.research.horizontal.HorizontalNoun;
 public class NounMine extends HorizontalNoun{
     @Override
-    public double getWidth(double size){
+    public float getWidth(float size){
         return size*3;
     }
     @Override
-    public double getHeight(double size){
+    public float getHeight(float size){
         return size;
     }
     @Override
-    protected void render(double x, double y, double w, double h){
+    protected void render(float x, float y, float w, float h){
         Core.drawHollowCosGear(x, y, h/4, h/2, h/50, 6, 0, 0, 3);//gear
         //right
         Core.drawOval(x+w/3, y, w/6, h/2, h/28, 100, 0, 58, 91);//small oval
