+++ src/planetaryprotector/research/horizontal/tense/TensePresent.java
@@ -3,7 +3,7 @@ import planetaryprotector.Core;
 import planetaryprotector.research.horizontal.HorizontalTense;
 public class TensePresent extends HorizontalTense{
     @Override
-    protected void render(double x, double y, double size){
+    protected void render(float x, float y, float size){
         Core.drawOval(x, y+size/4, size/4, size/4, size/22, 80, 0, -12, 12);//Bottom
         Core.drawOval(x, y-size/4, size/4, size/4, size/22, 80, 0, 28, 52);//Top
         Core.drawOval(x-size/4, y, size/4, size/4, size/22, 80, 0, 8, 32);//Left
